requirements.txt
celery==5.2.3
Django==4.0.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
psycopg2==2.9.3
redis==4.1.2
geoip2==4.5.0
pillow==9.0.0
requests~=2.27.1
asgiref~=3.5.0
kombu~=5.2.3
PyYAML~=6.0
PyJWT~=1.7.1
gunicorn==20.1.0
gevent
geocoder==1.38.1
django-redis==5.2.0
httpx==0.22.0
uvloop==0.16.0
uvicorn==0.17.6
pytrovich==0.0.2
docx-mailmerge2==0.6.2
qrcode==7.3.1
docxtpl==0.16.0